fixing event handlers and config
